requirements.txt
click
bs4
requests
tqdm